third_party/WebKit/LayoutTests/fast/css/foreignObject-bloom-filter.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>foreignObject #green { color: green }</style>
<foreignobject>
    <span id="green">This text should be green.</span>
</foreignobject>
<script>
description("Check that the selector filter does not reject a foreignobject selector due to camelCasing.");

shouldBeEqualToString("getComputedStyle(green).color", "rgb(0, 128, 0)");
</script>
